Bump svgparser to 0.6.
