+++ bench/svg_bench/Cargo.toml
@@ -11,7 +11,6 @@ path = "src/svg_renderer.rs"
 [dependencies]
 lyon = { path = "../../" }
 lyon_svg = { path = "../../svg" }
-svgparser = "0.4"
 glium = "0.15.0"
 glutin = "0.6.1"
 rayon = "0.4.2"
